refactor: auto-split test_analysis_comprehensive.py
